Added Highlight functionality to selecting towers for upgrades
Added Image
-----------------------
added hilite.png that uses convert_alpha() to show transparency

Edited main.py
-----------------------------
when tower is selected, a yellow box will surround it, making it easy for the player to tell which tower is being upgraded when the upgrade button is clicked. This required adding an extra parameter to the draw_window() function called "hilite".